code/requirements.txt
# ── Web framework ──────────────────────────────────────────────
Django>=5.1,<5.2

# ── PDF processing ─────────────────────────────────────────────
PyMuPDF>=1.24.0

# ── HTTP / scraping ────────────────────────────────────────────
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.2.0

# ── Environment config ─────────────────────────────────────────
python-decouple>=3.8

# ── LLM providers ─────────────────────────────────────────────
# Groq uses OpenAI-compatible endpoint; we point the openai SDK at it
openai>=1.30.0
# New unified Google GenAI SDK (google-generativeai is EOL as of Nov 2025)
google-genai>=1.0.0

# ── Data / repository crawling ─────────────────────────────────
kaggle>=2.0.0

# ── Utilities ─────────────────────────────────────────────────
Pillow>=10.3.0

# Vega-Altair chart rendering (custom NL chart pipeline)
altair>=5.3
vl-convert-python>=1.6
pandas>=2.0